Both public plans and project are bulk downloading now
